Move processPostTypeQuery to functions
