+++ functions/wordpress/postTypes/getHeadlessConfigPage.js
@@ -1,5 +1,5 @@
+import processPostTypeQuery from '@/functions/wordpress/postTypes/processPostTypeQuery'
 import headlessConfigPageQuerySeo from '@/lib/wordpress/_config/headlessConfigPageQuerySeo'
-import processPostTypeQuery from '@/lib/wordpress/_global/processPostTypeQuery'
 
 /**
  * Retrieve single page set via Headless Config.
