+++ functions/wordpress/postTypes/processPostTypeQuery.js
